Refactor the whole thing
