--- .github/workflows/_test-validate-inputs.yml
@@ -1,38 +1,0 @@
----
-name: Test validate-inputs action sub-workflow
-
-on:
-  workflow_call:
-    inputs:
-      number:
-        description: A number
-        required: true
-        type: string
-      alphanumeric:
-        description: An alphanumeric string
-        required: true
-        type: string
-      fruit:
-        description: A fruit
-        required: true
-        type: string
-
-jobs:
-  main:
-    name: Test validate-inputs action
-    runs-on: ubuntu-20.04
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v4
-
-      - name: Validate inputs
-        uses: ./validate-inputs
-        with:
-          names: |-
-            number
-            alphanumeric
-            fruit
-          regexps: |-
-            [0-9]+
-            [a-z0-9]+
-            apple|banana|lichi
