--- .github/workflows/test-validate-inputs.yml
@@ -1,46 +1,0 @@
----
-name: Test validate-inputs action
-
-on:
-  pull_request:
-    branches:
-      - main
-  push:
-    branches:
-      - main
-
-concurrency:
-  group: ${{ github.ref }}-${{ github.workflow }}
-  cancel-in-progress: true
-
-jobs:
-  check-valid-inputs:
-    name: Call sub-workflow with valid inputs
-    uses: ./.github/workflows/_test-validate-inputs.yml
-    with:
-      number: 44
-      alphanumeric: abc123
-      fruit: lichi
-
-  check-invalid-inputs:
-    needs: [check-valid-inputs]
-    name: Call sub-workflow with invalid inputs
-    uses: ./.github/workflows/_test-validate-inputs.yml
-    with:
-      number: a
-      alphanumeric: 34-a
-      fruit: lemon
-
-  check-failure:
-    needs: [check-invalid-inputs]
-    if: always()
-    name: Check failure
-    runs-on: ubuntu-20.04
-    steps:
-      - name: Check failure
-        shell: bash
-        run: |
-          if [[ "${{ needs.check-invalid-inputs.outcome }}" == "success" ]]; then
-            echo "!!! Failure test failed.  Something fishy is going on here"
-            exit 1
-          fi
