--- validate-inputs/action.yml
@@ -1,13 +1,0 @@
----
-name: Validate inputs
-description: This action validates inputs
-inputs:
-  names:
-    description: Names of inputs to be checked against regular expression. One per line
-    required: true
-  regexps:
-    description: Regular expressions to check against. One per line
-    required: true
-runs:
-  using: 'node20'
-  main: 'dist/index.js'
